FINERACT-2181: modified id and id related fields in api objects to long datatype

+++ fineract-e2e-tests-core/src/test/java/org/apache/fineract/test/factory/SavingsAccountRequestFactory.java
@@ -28,14 +28,14 @@ public final class SavingsAccountRequestFactory {
     public static final String DATE_FORMAT = "dd MMMM yyyy";
     public static final String DEFAULT_LOCALE = "en";
     public static final String DEFAULT_TRANSACTION_DATE = "";
-    public static final Integer DEFAULT_CLIENT_ID = 1;
+    public static final Long DEFAULT_CLIENT_ID = 1L;
     public static final String DEFAULT_SUBMITTED_ON_DATE = "";
     public static final String DEFAULT_APPROVED_ON_DATE = "";
     public static final String DEFAULT_ACTIVATED_ON_DATE = "";
     public static final BigDecimal DEFAULT_REPAYMENT_TRANSACTION_AMOUNT = new BigDecimal(1);
     public static final Integer DEFAULT_PAYMENT_TYPE_ID = 2;
-    public static final Integer EUR_SAVING_PRODUCT_ID = 1;
-    public static final Integer USD_SAVING_PRODUCT_ID = 2;
+    public static final Long EUR_SAVING_PRODUCT_ID = 1L;
+    public static final Long USD_SAVING_PRODUCT_ID = 2L;
 
     private SavingsAccountRequestFactory() {}
